# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.CitySummary:
    properties:
      aqi:
        type: integer
      aqi_name:
        type: string
      city:
        type: string
      county:
        type: string
      degree:
        type: string
      error:
        description: 错误信息
        type: string
      province:
        type: string
      tomorrow_max:
        type: string
      tomorrow_min:
        type: string
      updated_at:
        type: string
      weather:
        type: string
      weather_url:
        description: 天气的图标url
        type: string
    type: object
...
info:
  contact: {}
  description: Go-Web综合性应用API接口文档
  title: Go-Web项目 API
  version: 0.0.3
paths:
...
  /weather/top10:
    get:
      description: 并行获取10个热门城市的天气数据，包括温度、天气状况、空气质量等信息
      produces:
      - application/json
      responses:
        "200":
          description: 10个城市的天气数据数组
          schema:
            items:
              $ref: '#/definitions/controllers.CitySummary'
            type: array
        "401":
          description: 用户未授权
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 获取热门城市天气
      tags:
      - Weather
  /weather/user:
    get:
      description: 返回当前用户的名称、位置信息及对应城市的天气数据
      produces:
      - application/json
      responses:
        "200":
          description: 用户信息、位置和天气数据
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 用户未授权
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 获取失败
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 获取用户信息及天气
      tags:
      - Weather
swagger: "2.0"